+++ CHANGES.rst
@@ -5,6 +5,7 @@ Version 1.8
 -----------
 
 * Improve error handling when a command is not found.
+* Improve support for dotted options in pyproject.toml.
 
 
 Version 1.7