+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 1.8
 * Improve error handling when a command is not found.
 * Improve support for dotted options in pyproject.toml.
 * Load setup.cfg or pyproject.toml in scmver script.
+* Drop Python 3.8 support.
 
 
 Version 1.7